backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
websockets==12.0
pydantic==2.7.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
sqlalchemy==2.0.30
aiosqlite==0.20.0
loguru==0.7.2